Check animal for nil and supply '<silence>'  for species and sound if so.
This is a common solution to this problem.  We're saying:
If the object I got is Nil, I'll supply to values, otherwise I'll send messages to the object.

To illustrate how crazy this is, imagine that instead of getting an Animal or a Nil, we
instead sometimes get an Animal and other times get a Truck.  Truck and Animal have
completely different APIs.  We'd be up in arms if we were expected to deal with this
kind of random input, yet somehow we accept Animal and/or Nil as ok input.  No. No, no, no.

I'm gonna back this out in the next commit and follow a different path.